feat(superisland): 拆分超级岛浮窗列表模式相关管理类
拆分FloatingReplicaListModeManager、FloatingReplicaMappingManager、FloatingReplicaNotificationManager、FloatingReplicaWindowManager四个核心管理类，实现超级岛悬浮通知的列表模式、映射管理、通知控制和浮窗界面管理功能，完善悬浮通知的多通知切换、超时销毁、状态恢复等能力。
